Implementing the times method directly in the Money superclass
The implementation of the `times` method is practically the same in both
currency subclasses, so to avoid duplication we've moved it to the Money
superclass. However, it has required a few changes:
- Instead of returning the subclass, we'll now return a `Money` object. As
the equality comparison now uses the `currency` field, this change won't
imply any problem in our codebase (as the passing tests prove)
- As we now need to instantiate `Money` objects, the `Money` class has
stopped being an abstract class to become a regular one.