Creating the Expression class
The goal of the Expression interface is to declare the operations that
will be used to combine several types of currencies (e.g. addition).

For the moment we'll leave it empty for simplicity, but it will be used
soon and it's good that we have in mind the structure of the entity that
will represent combinations of different currencies (e.g. "$2 + 3CHF")