Implementing Sum.reduce
In order to avoid the bank having to access the Sum elements directly
(i.e. augend.amount and addend.amount), we'll implement a reduce element
in Sum and it's through it where the sum will be reduced to a Money.

For the moment we've focused on the base case, where both elements
belong to the same currency