Moving the reduce method to Expression
We need to implement reduce on Money too, even if it only returns
itself. Therefore, in order to make the implementation cleaner, we'll
convert it into an operation of the Expression interface, that both Sum
and Money implement.

Once more, we have only focused on single-currency operations. The
support for multiple currencies will come in later code changes