Enabling reduction to a different currency
This is the first step to enable multi-currency operations. We have
added a method (`rate`) to `Bank` that will return the change rate
between two different currencies

Of course the implementation is not complete (the rate is still hardcoded),
but we have started declarating the API that will be used.