Implementing rates correctly
Implementing the methods `addRate` and `rate` to add and retrieve rates
from a Bank respectively.

These method have been implemented using a Hashtable, in which the key
is a `Pair` representing a pair of currencies and the key is an integer
(the rate).

As Pair needs to be hashable to be used as a key, we have implemented
both the `equals` and the `hashCode` methods for it. However, we have
set the hashcode to always be 0 for simplicity, and because due to the
fact that we won't have much use-cases yet the performance decrease won't
be noticeable